refactor(compiler): separar ASTTranspiler em transpiladores por categoria (variables, functions, logics)
