+++ packages/compiler/src/index.ts
@@ -1,1 +1,4 @@
+export { ASTTranspiler } from "./ASTTranspiler.ts";
 export { VariablesTranspiler } from "./VariablesTranspiler.ts";
+export { FunctionsTranspiler } from "./FunctionsTranspiler.ts";
+export { LogicsTranspiler } from "./LogicsTranspiler.ts";
